bug(downsampler): Thread local memory constructs for multi-threaded spark executors (#510)
When there are multiple executors per box, each thread needs its own memory constructs. Earlier, downsampler job was sharing memory constructs among all executors on a worker - we now use thread locals for the off-heap memory.

Also worked around a possible bug in endMetaSpan.
